requirements.txt
Flask==1.0.2
gunicorn==19.6.0
mongoengine==0.17.0
flask-mako==0.4
flask-mongoengine==0.9.5
requests==2.11.1
-e git+https://github.com/dongweiming/fake-useragent.git@master#egg=fake-useragent
lxml==3.6.4
futures==3.0.5
pycrypto==2.6.1
redis==2.10.5
PyYAML==5.1